# myhomebp-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyHomeBP API",
        "description": "Blood Pressure Management API for NHS and Private GP Surgery Use in the UK",
        "contact": {
            "name": "247 Meditech Support",
            "email": "support@247meditech.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.myhomebp.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/reports/generate": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Generate blood pressure report",
                "operationId": "58f9d9409f09dee37db75ec7cf92a751",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-07"
                                    },
                                    "include_clinical_data": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "email_to_clinic": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Report generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Report generated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "report": {
                                                    "type": "object"
                                                },
                                                "pdf_url": {
                                                    "type": "string",
                                                    "example": "/storage/reports/report_123.pdf"
                                                },
                                                "email_sent": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/summary": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get report summary",
                "operationId": "78ee8a760db0d01c12d2963e87361f7f",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Number of days to include in summary",
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report summary retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "summary": {
                                                    "type": "object"
                                                },
                                                "can_generate_report": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "readings_count": {
                                                    "type": "integer",
                                                    "example": 14
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/download/{filename}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Download report PDF",
                "operationId": "f62187b60732dc90d1656da2a46dbeaf",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "Report filename",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "report_123.pdf"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report downloaded successfully"
                    },
                    "404": {
                        "description": "Report not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Report not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/history": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get report history",
                "operationId": "130dc212431d00a3a8fdda1ced6c7d41",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report history retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "reports": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token with the `Bearer: ` prefix, e.g. `Bearer abc123`",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}